@@ -11,20 +11,20 @@
         }
     ],
     "require": {
-        "php": "^5.3.3|^7.0",
-        "egeloen/json-builder": "^1.0|^2.0",
-        "symfony/dependency-injection": "^2.2|^3.0",
-        "symfony/form": "^2.2|^3.0",
-        "symfony/framework-bundle": "^2.2|^3.0"
+        "php": "^5.6|^7.0",
+        "egeloen/json-builder": "^2.0",
+        "symfony/dependency-injection": "^2.7|^3.0",
+        "symfony/form": "^2.7|^3.0",
+        "symfony/framework-bundle": "^2.7|^3.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.0",
-        "phpunit/phpunit": "^4.0|^5.0",
-        "symfony/asset": "^2.2|^3.0",
+        "phpunit/phpunit": "^5.0",
+        "symfony/asset": "^2.7|^3.0",
         "symfony/phpunit-bridge": "^2.7|^3.0",
-        "symfony/templating": "^2.2|^3.0",
-        "symfony/twig-bridge": "^2.2|^3.0",
-        "symfony/yaml": "^2.0.5|^3.0",
+        "symfony/templating": "^2.7|^3.0",
+        "symfony/twig-bridge": "^2.7|^3.0",
+        "symfony/yaml": "^2.7|^3.0",
         "twig/twig": "^1.12"
     },
     "suggest": {
