@@ -39,26 +39,26 @@ class ResourceCompilerPassTest extends AbstractTestCase
         $containerBuilder
             ->expects($this->exactly(2))
             ->method('hasParameter')
-            ->will($this->returnValueMap(array(
-                array('templating.helper.form.resources', false),
-                array($parameter = 'twig.form.resources', true),
-            )));
+            ->will($this->returnValueMap([
+                ['templating.helper.form.resources', false],
+                [$parameter = 'twig.form.resources', true],
+            ]));
 
         $containerBuilder
             ->expects($this->once())
             ->method('getParameter')
             ->with($this->identicalTo($parameter))
-            ->will($this->returnValue(array($template = 'layout.html.twig')));
+            ->will($this->returnValue([$template = 'layout.html.twig']));
 
         $containerBuilder
             ->expects($this->once())
             ->method('setParameter')
             ->with(
                 $this->identicalTo($parameter),
-                $this->identicalTo(array(
+                $this->identicalTo([
                     'IvoryCKEditorBundle:Form:ckeditor_widget.html.twig',
                     $template,
-                ))
+                ])
             );
 
         $this->compilerPass->process($containerBuilder);
@@ -70,23 +70,23 @@ class ResourceCompilerPassTest extends AbstractTestCase
         $containerBuilder
             ->expects($this->exactly(2))
             ->method('hasParameter')
-            ->will($this->returnValueMap(array(
-                array($parameter = 'templating.helper.form.resources', true),
-                array('twig.form.resources', false),
-            )));
+            ->will($this->returnValueMap([
+                [$parameter = 'templating.helper.form.resources', true],
+                ['twig.form.resources', false],
+            ]));
 
         $containerBuilder
             ->expects($this->once())
             ->method('getParameter')
             ->with($this->identicalTo($parameter))
-            ->will($this->returnValue(array($template = 'layout.html.php')));
+            ->will($this->returnValue([$template = 'layout.html.php']));
 
         $containerBuilder
             ->expects($this->once())
             ->method('setParameter')
             ->with(
                 $this->identicalTo($parameter),
-                $this->identicalTo(array('IvoryCKEditorBundle:Form', $template))
+                $this->identicalTo(['IvoryCKEditorBundle:Form', $template])
             );
 
         $this->compilerPass->process($containerBuilder);
@@ -97,9 +97,9 @@ class ResourceCompilerPassTest extends AbstractTestCase
      */
     private function createContainerBuilderMock()
     {
-        return $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerBuilder')
+        return $this->getMockBuilder(ContainerBuilder::class)
             ->disableOriginalConstructor()
-            ->setMethods(array('hasParameter', 'getParameter', 'setParameter'))
+            ->setMethods(['hasParameter', 'getParameter', 'setParameter'])
             ->getMock();
     }
 }
